LayoutTests/inspector/canvas/recording-offscreen-webgl-frameCount-expected.txt

Test that CanvasManager is able to record actions made to offscreen WebGL canvas contexts.


== Running test suite: Canvas.recordingOffscreenWebGL
-- Running test case: Canvas.recordingOffscreenWebGL.singleFrame
initialState:
  attributes:
    width: 2
    height: 2
  parameters:
    0: {"alpha":true,"depth":true,"stencil":false,"antialias":true,"premultipliedAlpha":true,"preserveDrawingBuffer":false,"powerPreference":"default","failIfMajorPerformanceCaveat":false}
  content: <filtered>
frames:
  0: (duration)
    0: activeTexture(1)
      swizzleTypes: [Number]
      trace:
        0: activeTexture
        1: (anonymous function)
        2: executeFrameFunction
        3: performActions
        4: Global Code
